支持PD分离在开启Prefix Cache时跳过部分block的传输 (merge request !1296)
Squash merge branch 'pdprefix' into 'master'
PD分离提高命中Prefix Cache时的传输效率。DECODE节点告知PREFILL节点命中cache的block，PREFILL节点不传输此类block



 



TAPD: --story=126360319

+++ src/ksana_llm/runtime/infer_request.h
@@ -188,7 +188,7 @@ class InferRequest {
   int step = 0;
 
   // The number of tokens for which kv caches have been generated.
-  int kv_cached_token_num = 0;
+  size_t kv_cached_token_num = 0;
 
   size_t mtp_kv_cached_token_num = 0;
